% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allvertices.R
\name{allvertices}
\alias{allvertices}
\title{Vertices of an hypercube}
\usage{
allvertices(d, lower = NULL, upper = NULL)
}
\arguments{
\item{d}{dimension of the hypercube}

\item{lower}{Array of size d containing the lower bound of the domain in each dimension.}

\item{upper}{Array of size d containing the upper bound of the domain in each dimension.}
}
\value{
A matrix with 2^d rows and d columns containing all the vertices.
}
\description{
Builds the 2^d vertices of an hypercube
}
\examples{
allvertices(d=3,lower=c(0,-10,10),upper=c(100,0,1000))
}
\author{
Clement Chevalier \email{clement.chevalier@unine.ch}
}

