% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/b2_l_kinf_cx_hx-data.R
\docType{data}
\name{b2_l_kinf_cx_hx}
\alias{b2_l_kinf_cx_hx}
\title{Dataset for computing keff}
\format{A matrix}
\usage{
b2_l_kinf_cx_hx
}
\description{
Dataset for computing keff
}
\author{
Gregory Caplin
2015-10-01
}

