% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branin_robinv.R
\name{branin_robinv}
\alias{branin_robinv}
\title{Branin function with additional nuisance parameters}
\usage{
branin_robinv(x)
}
\arguments{
\item{x}{vector of size 3 or more, or matrix with 3 columns or more.}
}
\value{
a scalar
}
\description{
Similar to the branin function of the DiceKriging package, but with additional
mute variables, serving as nuisance parameters.
}
\examples{
library(DiceKriging)

branin_robinv(c(0.5,0.4,0.6))

xx <- matrix(runif(50),nrow=10)
branin_robinv(xx)
}
\author{
Clement Chevalier \email{clement.chevalier@unine.ch}
}
