% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EGRI.R
\name{EGRI}
\alias{EGRI}
\title{Efficient Global Robust Inversion}
\usage{
EGRI(T, model, method = NULL, fun, iter = 1, batchsize = 1, opt.index,
  inv.index, lower = NULL, upper = NULL, new.noise.var = NULL,
  integcontrol = NULL, optimcontrol = NULL, kmcontrol = NULL, ...)
}
\arguments{
\item{T}{Target threshold.}

\item{model}{The current kriging model. km object.}

\item{method}{Criterion used for choosing observations.
Currently, only the \code{"surnew"} criterion is available.}

\item{fun}{Objective function.}

\item{iter}{Number of iterations. At each iteration, a batch of batchsize point is evaluated in parallel.}

\item{batchsize}{The size of the batch of points evaluated at each iteration.}

\item{opt.index}{Array with integers corresponding to the indices of the nuisance parameters.}
  
\item{inv.indev}{Array with integers corresponding to the indices of the controlled parameters.}

\item{lower}{Array of size d. Lower bound of the input domain.}

\item{upper}{Array of size d. Upper bound of the input domain.}

\item{new.noise.var}{Optional scalar value of the noise variance of the new observations.}

\item{integcontrol}{The integcontrol argument used in the \code{integration_design_robinv} function.
This applies for the following criterion: \code{"surnew"}. See the help of the
\code{integration_design_robinv} function.}

\item{optimcontrol}{A list used to set up the optimization procedure of the chosen sampling criterion.
For the \code{"surnew"} criterion, see the help of the \code{max_surnew_robinv} function.}

\item{kmcontrol}{Optional list representing the control variables for the re-estimation of the kriging model once new points are sampled.
The items are the same as in the \code{km} function of the DiceKriging package.}

\item{...}{Other arguments of the objective function fun.}
}
\value{
A list with the following fields.
(i) par: The added observations (iter*batchsize) x d matrix,
(ii) value: The value of the function fun at the added observations,
(iii) lastmodel: The current (last) kriging model of km class,
(iv) lastvalue: The value of the criterion at the last added point,
(v) allvalues: If an optimization on a discrete set of points is chosen, the value of the criterion at all these points, for the last iteration,
(vi) lastintegration.param: For debug. The last value returned by the integration_design_robinv function, if applicable.
}
\description{
Sequential design of experiments for robust inversion.
}
\examples{
library(KrigInv)
myfun <- function(x) return(-1 * branin_robinv(x))
d <- 4

set.seed(8)

# an example with scaling
n0 <- 40
T <- -10
opt.index <- c(3,4)
inv.index <- c(1,2)
lower <- rep(0,times=d)
upper <- rep(1,times=d)

design <- matrix(runif(d*n0),nrow=n0)
response <- myfun(design)

knots.number <- c(2,2,2,2)
knots <- generate_knots(knots.number = knots.number , d = d)

model <- km(formula = ~1,design = design,response = response,covtype = "matern3_2",scaling = TRUE,knots = knots)

integcontrol <- list(distrib = "surnew",n.points = 100,finaldistrib="surnew",n.candidates=300,
                     nsimu=100,n.optpoints = 50,choose_optpoints=TRUE,n.optpoints.candidates=200)

optimcontrol <- list(method = "genoud", pop.size = 400, max.generations = 4, wait.generation = 1)
\dontrun{
obj <- EGRI(T = T,method="surnew",model = model,fun = myfun,iter = 2,
            batchsize = 2,opt.index = opt.index,inv.index = inv.index,
            integcontrol=integcontrol,optimcontrol=optimcontrol)
}
}
\author{
Clement Chevalier \email{clement.chevalier@unine.ch}
}

