% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_knots.R
\name{generate_knots}
\alias{generate_knots}
\title{Generating the knots used in a km call}
\usage{
generate_knots(knots.number = NULL, d, lower = NULL, upper = NULL)
}
\arguments{
\item{knots.number}{Array of size d containing integers. 
\code{knots.number[i]} contains the number of knots in dimension i. Values of 1 or 0 are considered as 
no knots.}

\item{d}{Integer. Number of dimensions. This is also equal to \code{length(knots.number)}}

\item{lower}{Array of size d containing the lower bounds of the input domain.}

\item{upper}{Array of size d containing the upper bounds of the input domain.}
}
\value{
A list with d fields. Field i is an array of size \code{knots.number[i]} when there is 2 knots 
or more, and is 0 otherwise.
}
\description{
Generates automatically the object used as an argument of the km function of the DiceKriging package
}
\details{
The knots are placed at the boundary of the domain, so, if we work in the unit hypercube then 
the array \code{result[[i]]} will be \code{c(0,1)} if  \code{knots.number[i] = 2}, 
\code{c(0,0.5,1)} if  \code{knots.number[i] = 3}, \code{c(0,1/3,2/3,1)} if  \code{knots.number[i] = 4} 
and so on.
}
\examples{

knots <- c(5,3,1,0,2,4)

result <- generate_knots(knots.number=knots,d=6)

}
\author{
Clement Chevalier \email{clement.chevalier@unine.ch}
}
