% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keff.R
\name{keff}
\alias{keff}
\title{3d test function computing a k-effective in nuclear safety}
\usage{
keff(x)
}
\arguments{
\item{x}{Array of size 3 or matrix with 3 columns. All input variables are 
in the interval [0,1].}
}
\value{
If x is a matrix with n rows and 3 columns, the 
function is evaluated on each row and returns an array of size n. If x is an array, 
the function returns a scalar (the k-effective).
}
\description{
3d test function computing a k-effective in nuclear safety
}
\details{
This function is important in robust inversion since the first parameter 
of the function is considered as a nuisance parameter while the second and third 
parameters are controlled.
}
\examples{
keff(runif(3))
}
\author{
Gregory Caplin \email{gregory.caplin@irsn.fr}
}
