% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCol.R
\name{plotCol}
\alias{plotCol}
\title{Comparing Colors}
\usage{
plotCol(col, nrow = 1, ncol = ceiling(length(col)/nrow),
  txt.col = "black")
}
\arguments{
\item{col}{Array of colors to plot}

\item{nrow}{Number of rows of the plot}

\item{ncol}{Number of columns of the plot}

\item{txt.col}{Some text to insert}
}
\value{
the grid layout, invisibly
}
\description{
Comparing Colors
}
\author{
Marius Hofert, originally
}
