% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translude.R
\name{translude}
\alias{translude}
\title{Using transparence to modify colors}
\usage{
translude(colors, alpha = 0.6)
}
\arguments{
\item{colors}{A color or an array where each element is a color.}

\item{alpha}{Scalar taking values between 0 (fully transparent) and 1 (opaque).}
}
\value{
An array with the colors transluded with opacity alpha
}
\description{
Modifies a color by playing on its transparence. Low alpha make the color more transparent 
while larger alpha make it opaque.
}
\examples{
library("grid")
onecolor <- "blue"

shades.count <- 10

alpha_tab <- seq(from = 0, to = 1, length=shades.count)
result <- rep("",times = shades.count)

for(i in 1:shades.count) result[i] <- translude(colors=onecolor,alpha_tab[i])
plotCol(result,txt.col=result)
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
